backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
